Factor out balance out of book_keeping.
