+++ core/src/report/book_keeping.rs
@@ -1,7 +1,7 @@
 //! Contains book keeping logics to process the input stream,
 //! and convert them into a processed Transactions.
 
-use std::{borrow::Borrow, collections::HashMap};
+use std::borrow::Borrow;
 
 use bumpalo::collections as bcc;
 use chrono::NaiveDate;
@@ -9,6 +9,7 @@ use chrono::NaiveDate;
 use crate::{load, repl};
 
 use super::{
+    balance::Balance,
     context::{Account, ReportContext},
     error::{self, ReportError},
     eval::{Amount, EvalError, Evaluable},
@@ -135,31 +136,6 @@ pub struct Posting<'ctx> {
     pub amount: Amount<'ctx>,
 }
 
-/// Balance of all accounts after the accumulated transactions.
-#[derive(Debug, Default)]
-pub struct Balance<'ctx> {
-    accounts: HashMap<Account<'ctx>, Amount<'ctx>>,
-}
-
-impl<'ctx> Balance<'ctx> {
-    /// Adds a particular account value, and returns the updated balance.
-    pub fn increment(&mut self, account: Account<'ctx>, amount: Amount<'ctx>) -> Amount<'ctx> {
-        let curr: &mut Amount = self.accounts.entry(account).or_default();
-        *curr += amount;
-        curr.clone()
-    }
-
-    /// Sets the particular account's balance, and returns the previous balance.
-    pub fn set_balance(&mut self, account: Account<'ctx>, amount: Amount<'ctx>) -> Amount<'ctx> {
-        self.accounts.insert(account, amount).unwrap_or_default()
-    }
-
-    /// Gets the balance of the given account.
-    pub fn get_balance(&self, account: &Account<'ctx>) -> Option<&Amount<'ctx>> {
-        self.accounts.get(account)
-    }
-}
-
 /// Adds a repl transaction, and converts it into a processed Transaction.
 fn add_transaction<'ctx>(
     ctx: &mut ReportContext<'ctx>,
@@ -218,92 +194,4 @@ fn add_transaction<'ctx>(
         date: txn.date,
         postings: postings.into_boxed_slice(),
     })
-}
-
-#[cfg(test)]
-mod tests {
-    use bumpalo::Bump;
-    use pretty_assertions::assert_eq;
-    use rust_decimal_macros::dec;
-
-    use super::*;
-
-    #[test]
-    fn balance_gives_zero_amount_when_not_initalized() {
-        let arena = Bump::new();
-        let mut ctx = ReportContext::new(&arena);
-
-        let balance = Balance::default();
-        assert_eq!(balance.get_balance(&ctx.accounts.ensure("Expenses")), None);
-    }
-
-    #[test]
-    fn test_balance_increment_adds_value() {
-        let arena = Bump::new();
-        let mut ctx = ReportContext::new(&arena);
-
-        let mut balance = Balance::default();
-        let updated = balance.increment(
-            ctx.accounts.ensure("Expenses"),
-            Amount::from_value(dec!(1000), ctx.commodities.ensure("JPY")),
-        );
-
-        assert_eq!(
-            updated,
-            Amount::from_value(dec!(1000), ctx.commodities.ensure("JPY"))
-        );
-        assert_eq!(
-            balance.get_balance(&ctx.accounts.ensure("Expenses")),
-            Some(&updated)
-        );
-
-        let updated = balance.increment(
-            ctx.accounts.ensure("Expenses"),
-            Amount::from_value(dec!(-1000), ctx.commodities.ensure("JPY")),
-        );
-
-        assert_eq!(updated, Amount::zero());
-        assert_eq!(
-            balance.get_balance(&ctx.accounts.ensure("Expenses")),
-            Some(&updated)
-        );
-    }
-
-    #[test]
-    fn test_balance_set_balance() {
-        let arena = Bump::new();
-        let mut ctx = ReportContext::new(&arena);
-
-        let mut balance = Balance::default();
-        let prev = balance.set_balance(
-            ctx.accounts.ensure("Expenses"),
-            Amount::from_value(dec!(1000), ctx.commodities.ensure("JPY")),
-        );
-
-        assert_eq!(prev, Amount::zero());
-        assert_eq!(
-            balance.get_balance(&ctx.accounts.ensure("Expenses")),
-            Some(&Amount::from_value(
-                dec!(1000),
-                ctx.commodities.ensure("JPY")
-            ))
-        );
-
-        let prev = balance.set_balance(
-            ctx.accounts.ensure("Expenses"),
-            Amount::from_value(dec!(-1000), ctx.commodities.ensure("JPY")),
-        );
-
-        assert_eq!(
-            prev,
-            Amount::from_value(dec!(1000), ctx.commodities.ensure("JPY"))
-        );
-        assert_eq!(
-            balance.get_balance(&ctx.accounts.ensure("Expenses")),
-            Some(&Amount::from_value(
-                dec!(-1000),
-                ctx.commodities.ensure("JPY")
-            ))
-        );
-    }
 }
